+++ livekit-datatrack/CHANGELOG.md
@@ -1,1 +1,30 @@
 # Changelog
+## 0.1.5 (2026-05-11)
+
+### Fixes
+
+- Upgrade protocol to v1.45.8
+
+## 0.1.4 (2026-04-23)
+
+### Fixes
+
+- Fix data track packet format issue breaking E2EE
+
+## 0.1.3 (2026-04-02)
+
+### Features
+
+- Rename type to `DataTrackStream`
+
+## 0.1.2 (2026-03-31)
+
+### Fixes
+
+- Upgrade to thiserror 2
+
+## 0.1.1 (2026-03-22)
+
+### Features
+
+- Initial release.